+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- --preserve-fields option to the cli so the user can specify which fields to keep when flattening, instead of defaulting to all of them
+
 ## [0.7.0] - 2019-04-08
 
 ## Fixed
@@ -31,7 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- While trying to decode JSON, if there is a UTF8 error raise the new exception class BadlyFormedJSONErrorUTF8. 
+- While trying to decode JSON, if there is a UTF8 error raise the new exception class BadlyFormedJSONErrorUTF8.
   (This extends the old BadlyFormedJSONError exception, so existing code that checks sensibly should be fine.)
 
 ## [0.4.0] - 2018-11-07