+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Correct cell source map when hashcomments is used.
 
+### Changed
+
+- Using --rollup with `flatten` accepts fields to roll up as input directly and via a file, as well as via a schema.
+
 ## [0.7.0] - 2019-04-08
 
 ## Fixed